backend/requirements.txt
Django>=6.0,<6.1
djangorestframework
markdown       # Útil para a API navegável do DRF
django-filter  # Para filtros avançados na API
psycopg2-binary
django-environ # Para ler o arquivo .env
django-redis
gunicorn       # Servidor de produção (já deixamos pronto)
pytest-django  # Para nossos testes TDD
django-cors-headers
djangorestframework-simplejwt
Pillow